CHROMIUM: iwl7000: Use timer_shutdown*() where appropriate
Due to several bugs caused by timers being re-armed after they are
shutdown and just before they are freed, a new state of timers was added
called "shutdown". After a timer is set to this state, then it can no
longer be re-armed, and will trigger a warning if it is.

The following script was run to find all the trivial locations in the
iwl7000 code where del_timer() or del_timer_sync() is called in the same
function that the object holding the timer is freed. It also ignores any
locations where the timer->function is modified between the del_timer*()
and the free(), as that is not considered a "trivial" case.

The coccinelle script:

    @@
    expression E, ptr, slab;
    identifier timer, rfield;
    @@
    (
    -       del_timer(&ptr->timer);
    +       timer_shutdown(&ptr->timer);
    |
    -       del_timer_sync(&ptr->timer);
    +       timer_shutdown_sync(&ptr->timer);
    )
      ... when strict
          when != ptr->timer.function = E;
    (
            kfree_rcu(ptr, rfield);
    |
            kmem_cache_free(slab, ptr);
    |
            kfree(ptr);
    )

UPSTREAM-TASK=b:258111198
BUG=b:222168464
TEST=CQ

Change-Id: Ib49335b70d630611fa5f5ac28832854ec1d91fe0
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4013625
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>