CHROMIUM: iwl7000: mvm: offload BSS Transition Management response during D3
There are mainly two types of BTM requests, recommendations and
notifications. For the first type, a response is needed otherwise,
most probably the STA will be disconnected.
Since we don't want to wake up the host on it set the BTM to reject offload
flag (if the device supports it) and rely on the FW to take care of it.
The FW will reject the BTM request and in case the AP sends DEAUTH
the FW can wake up the host to let it decide on the next steps.

Change-Id: I9457acc55bc23ce715c714b5088058f52540c224
Signed-off-by: Haim Dreyfuss <haim.dreyfuss@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 5d90d1d6d56ef6c9fc9c8894c6c544e687ccda06