CHROMIUM: Do not use timer namespace for timer_shutdown() function
A new "shutdown" timer state is being added to the generic timer code. One
of the functions to change the timer into the state is called
"timer_shutdown()". This means that there can not be other functions
called "timer_shutdown()" as the timer code owns the "timer_*" name space.

Rename timer_shutdown() to <driver>arch_timer_shutdown() to avoid this
conflict.

This patch is only necessary in chromeos-4.19 since the affected drivers
will either be modified or dropped in later ChromeOS branches.

UPSTREAM-TASK=b:258111198
BUG=b:222168464
TEST=CQ

Change-Id: I9e779fa71f0d869993db36820cf93a9d768f44d2
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4131877
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>