CHROMIUM: iwl7000: add some FW misbehaviour check infrastructure
When the firmware misbehaves (according to the driver), we
often either ignore that, or WARN_ON, which is very noisy
but doesn't really help.

Add a little helper macro IWL_FW_CHECK() that can be used
in place of WARN_ON() in conditions, and make it take a
message that's printed in this case. We can also add more
behaviour to this in the future.

Use it in a single place that we've encountered recently,
but that's more to have an example; there are a lot more
places that should use it.

Change-Id: Ia0198036b7a626876d836bd41a4b2d2b1e65c5ca
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: d79f6fc28b982ee69a0a1a8b1cce2d298f7a9273