CHROMIUM: iwl7000: ieee80211: use default for medium synchronization delay
Default values are defined for the information included in the Medium
Synchronization Delay Information subfield. The spec says to
initialize the values to these defaults and only change them when
included.

Return the default value instead of zero so that the defaults are
used when the field is not included in the association response.

Change-Id: I2d3528cf4af021c5b37f97fbe64ae9116ce9bef1
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: fef0c9ac58221555f25171d6ad235dd5043b1d00