FROMGIT: Bluetooth: hci_sync: Resume adv with no RPA when active scan
The address resolution should be disabled during the active scan,
so all the advertisements can reach the host. The advertising
has to be paused before disabling the address resolution,
because the advertising will prevent any changes to the resolving
list and the address resolution status. Skipping this will cause
the hci error and the discovery failure.

According to the bluetooth specification:
"7.8.44 LE Set Address Resolution Enable command

This command shall not be used when:
- Advertising (other than periodic advertising) is enabled,
- Scanning is enabled, or
- an HCI_LE_Create_Connection, HCI_LE_Extended_Create_Connection, or
  HCI_LE_Periodic_Advertising_Create_Sync command is outstanding."

If the host is using RPA, the controller needs to generate RPA for
the advertising, so the advertising must remain paused during the
active scan.

If the host is not using RPA, the advertising can be resumed after
disabling the address resolution.

Fixes: 9afc675edeeb ("Bluetooth: hci_sync: allow advertise when scan without RPA")
Signed-off-by: Zhengping Jiang <jiangzp@google.com>
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
(cherry picked from commit 2b291083eb0c49cd1633cddf95cc0e6e7303e43b
 git://git.kernel.org/pub/scm/linux/kernel/git/bluetooth/bluetooth-next.git/ master)

BUG=b:267966317
TEST=turn on nearby share and LL privacy, scan from UI
TEST=bluetooth_AdapterLLPrivacyHealth.le_role_sender

Change-Id: Ibe4d3a42683381c1e78b8c3aa67b53fc74437ae9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4296707
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Zhengping Jiang <jiangzp@google.com>
Tested-by: Zhengping Jiang <jiangzp@google.com>