CHROMIUM: iwl7000: mvm: fix radiotap RU allocation reporting
The RU allocation fields are reported in the wrong order,
fix that according to what the hardware/firmware is saying
the order should be.

Also extend the macros to have a lookup of the value to
its position in the data fields for both radiotap and FW,
to simplify the actual code.

Change-Id: Ia33f29da9442ff046773896ba43a7186cdde95b9
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 8338891c70ac5670a3e5dfbeb4cecb5d841877a5