BACKPORT: media: vivid: drop overlay support
Destructive overlay support (i.e. where the video frame is DMA-ed
straight into a framebuffer) is effectively dead. It was a
necessary evil in the early days when computers were not fast enough
to copy SDTV video frames around, but today that's no longer a problem.

It requires access to the framebuffer memory, which is a bad idea and
very hard to do safely. In addition, in drm it is today almost
impossible to get hold of the framebuffer address.

So drop support for this.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
(cherry picked from commit ccaa9d50ca73d136accbf66798c04fb25ca44bd5)

Conflicts:
   drivers/media/test-drivers/vivid/vivid-vid-cap.c

BUG=b:380510063
TEST=v4l2-compliance vivid

Change-Id: Ic6dd953757c50f2522b015322a7702d3449b37c5
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6063143
Reviewed-by: Yunke Cao <yunkec@chromium.org>
Commit-Queue: Yunke Cao <yunkec@chromium.org>
Tested-by: Yunke Cao <yunkec@chromium.org>