CHROMIUM: iwl7000: mac80211: cleanup auth_data only if association continues
If the association command fails then the authentication is still valid
and it makes sense to keep it alive. Otherwise, we would currently get
into an inconsistent state because mac80211 on the one hand is
disconnected but on the other hand the state is not entirely cleared
and a new authentication could not continue.

Change-Id: I7f3dcd4120a186484a91b87560e9b7201d40984f
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 53d05352d646b5fb36a90d7637e3047997e0187e