CHROMIUM: add cros_net tp for inet send,rcv,accept
!!!DO NOT UPSTREAM!!!
!!!DO NOT PORT TO OTHER KERNELS!!!

Add tracepoints that capture the exits of:
	inet_send
	inet_receive
	inet_stream_accept

The rationale is the same as the rationale used in review 4429631:
This patch is essentially the same as the patch in review 4571256
with the only difference being that this is for the 5.10 kernel.

That is, fentry/fexit bpf hooks for ARM64 are not available
in this kernel. Instead of cherry-picking many, many patches to support
ARM64 fentry/fexit hooks, we add these tracepoints.

These tracepoints sole purpose is to provide tracepoints for network
monitoring bpfs for the generating of network XDR events.

It is fully expected that these changes will be dropped going forward in
6.1+ kernels since fentry/fexit bpf attach points for ARM64 are expected
to be fully supported in newer kernels.

BUG=b:251499575
TEST=compiled and tested on arm64-generic, attached bpf to the
tracepoint and verified correct information was being received in
userspace.
UPSTREAM-TASK=b:278303728

Change-Id: I49ca192c99d13dcf66bbc2ae8ff2819c96e1b77a
Signed-off-by: Jason Ling <jasonling@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4582280
Reviewed-by: Steven Rostedt <rostedt@google.com>
Reviewed-by: Aashay Shringarpure <aashay@google.com>