FROMLIST: mm/khugepaged: fix iteration in collapse_file
Remove an unnecessary call to xas_set(index) when iterating over the
target range in collapse_file. The extra call to xas_set reset the xas
cursor to the top of the tree, causing the xas_next call on the next
iteration to walk the tree to index instead of advancing to index+1.
This returned the same page again, which would cause collapse_file to
fail because the page is already locked.

This bug was hidden when CONFIG_DEBUG_VM was set. When that config was
used, the xas_load in a subsequent VM_BUG_ON assert would walk xas from
the top of the tree to index, causing the xas_next call on the next loop
iteration to advance the cursor as expected.

Fixes: a2e17cc2efc7 ("mm/khugepaged: maintain page cache uptodate flag")
Signed-off-by: David Stevens <stevensd@chromium.org>
(am from https://patchwork.kernel.org/patch/13270002/)
(also found at https://lore.kernel.org/r/20230607053135.2087354-1-stevensd@google.com)

UPSTREAM-TASK=b:279553168
BUG=b:285053764
TEST=Manual verification with boosted khugepaged frequency

Change-Id: I00fcaf21e5f7659aec0355dc6cf69248919a991b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4610263
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>