BACKPORT: UPSTREAM: media: mediatek: vcodec: Read HW active status from syscon
Remove the requirement of a VDEC_SYS reg iospace for both MT8173 and
MT8183. To achieve that, rely on a vdecsys syscon to be passed through
the DT, and use it to directly read the VDEC_HW_ACTIVE bit during IRQ
handling to check whether the HW is active. Also update the VP8 stateful
decoder to use the syscon, if present, for writes to VDEC_SYS.

The old behavior is still present when reg-names aren't supplied, as
to keep backward compatibility.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
[hverkuil: added vdecsys_regmap kerneldoc line provided by Nicolas]
(cherry picked from commit c33262171252d355f9c962d603425d445a24ac4c)

BUG=b:260600757
TEST=emerge-cherry sys-kernel/chromeos-kernel-5_10

Change-Id: I3a27ddc1e3eb9ec228a058817c6a182837fc79dc
Signed-off-by: mingjia.zhang <mingjia.zhang@mediatek.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4970463
Reviewed-by: Steve Cho <stevecho@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>