CHROMIUM: cfg80211: make nl80211_attrs order align with the ones in the upstream kernel
The nl80211_attrs enum order does not match the one used in upstream
kernel when backported in
  commit eac3f2571ac5f1422f7c1df3db0b10417367f632

This mismatch causes troubles when userspace applications like
wpa_supplicant uses a nl80211_copy.h file which aligns with the upstream
kernel but not the one used in Chromium kernel. It will send mismatched
attibute values to kernel and cause error.

Instead of reverting and relanding the above commits and introduces
several conflicts in both reverting and relanding which is prone to
errors. This CL is the lesser of two evils and fix the ordering in a
CHROMIUM patch.

Signed-off-by: Jintao Lin <jintaolin@chromium.org>

BUG=b:259239952
TEST=build kernel

Change-Id: I21be03f3a27cbc30ec4f6a77331132eddbb1416d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4749308
Commit-Queue: Jintao Lin <jintaolin@chromium.org>
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jacek Siuda <jsiuda@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>