CHROMIUM: iwl7000: mvm: fix a race between the end and start of a roc
When we first inform mac80211 that an roc is expired and then schedule
the worker to cleanup everything in the driver, we can have the
following race.

 CPU0						CPU1
iwl_mvm_rx_session_protect_notif
  ieee80211_remain_on_channel_expired
  						ieee80211_hw_roc_done
						  ieee80211_end_finished_rocs
						   <roc_list is now empty>

						nl80211_tx_mgmt
						  <snip>
						  ieee80211_start_roc_work
						    <snip>
						    iwl_mvm_roc_common
						      flush_work(&mvm->roc_done_wk); (not scheduled yet)
						      // continue to add the roc to the FW
  iwl_mvm_p2p_roc_finished
    iwl_mvm_roc_finished
      schedule_work(&mvm->roc_done_wk);

At this stage, we have an roc that was added to the firmware and
roc_done_wk is scheduled. This is not a good idea. Then,

iwl_mvm_roc_done_wk
  iwl_mvm_link_changed, LINK_CONTEXT_MODIFY_ACTIVE, FALSE

This deactivates the link and since a session protection is active, the
firmware asserts with the code 320F.

Fix this by first calling iwl_mvm_p2p_roc_finished to schedule the
worker and only then inform mac80211 that the roc is expired. This way,
any new roc being added to iwlmvm will see that the roc_done worker
needs to complete first.

Change-Id: Iba16aa7eb0d3fc447bfb29a51da667d8cc33cdef
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: fbc31b4ca489fbf97c266f0cd13ed3a87100ad9a