FROMLIST: rtc: Add support for limited alarm timer offsets
Some alarm timers are based on time offsets, not on absolute times.
In some situations, the amount of time that can be scheduled in the
future is limited. This may result in a refusal to suspend the system,
causing substantial battery drain.

Some RTC alarm drivers remedy the situation by setting the alarm time
to the maximum supported time if a request for an out-of-range timeout
is made. This is not really desirable since it may result in unexpected
early wakeups.

To reduce the impact of this problem, let RTC drivers report the maximum
supported alarm timer offset. The code setting alarm timers can then
decide if it wants to reject setting alarm timers to a larger value, if it
wants to implement recurring alarms until the actually requested alarm
time is met, or if it wants to accept the limited alarm time.

Only introduce the necessary variable into struct rtc_device.
Code to set and use the variable will follow with subsequent patches.

Cc: Brian Norris <briannorris@chromium.org>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
(am from https://patchwork.ozlabs.org/patch/1822562/)
(also found at https://lore.kernel.org/r/20230817225537.4053865-2-linux@roeck-us.net)

UPSTREAM-TASK=b:257330044
BUG=b:254247476, b:291895945
TEST=Run on affected Chromebook

Change-Id: I40f399580cd9e2f3b2215bd8e3692f62380d7d27
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4004305
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Eizan Miyamoto <eizan@chromium.org>
Reviewed-by: Eizan Miyamoto <eizan@chromium.org>