FIXUP: CHROMIUM: tick-sched: Let stop code round to next tick using hrtimer_forward()
It is possible that when we stop the tick, the sched timer is not programmed because hrtimer_forward() doesn't do anything. To be safe, do
a hrtimer_set_expires() so that the hrtimer is setup to at least 'last_tick'. This fixes lag issues seen with a bluetooth mouse pointer lag.

BUG=b:297591899
TEST=Connect bluetooth mouse to Brya device, no pointer lag observed.
Change-Id: I549feae270c7883a9290f8284ab7a2aae827b5fc
Signed-off-by: Joel Fernandes <joelaf@google.com>
Signed-off-by: Suleiman Souhlal <suleiman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4892327
Reviewed-by: Vineeth Pillai <vineethrp@google.com>