CHROMIUM: iwl7000: mac80211: ibss: restore STA allocation on data RX
Revert the station info allocation in mac80211 back to the
upstream version. Back when we posted this, an issue with
this was noted in that it can cause us to send a lot of
probe request frames if we hear a lot of traffic from a
given station, and we never fixed this. Also, the premise
of this seems wrong, the original commit talks about how
iwlwifi cannot do atomic STA allocations, but as far as
the driver is concerned, the station is only added when
the incomplete station list is processed, non-atomically
later. TX to the station can also only happen after that.

Neither of the approaches really is ideal, we might want
to fix this properly, but until then we shouldn't have it
diverge from upstream and be able to cause some kind of
traffic amplification.

Change-Id: Iff818a1fb2b2e905a633dc555c6a300dd91cb92d
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 222c0852e6761745f14bc1c49c21e76b8ed685a5