UPSTREAM: media: ccs: Drop re-entrant s_stream support
The s_stream is called to enable and to disable streaming on a sub-device.
The caller may only call it to change the state, enabling streaming is not
allowed when it is already disabled, and similarly for disabling
streaming. Remove the check from the CCS driver.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
(cherry picked from commit 8a695a29d68e8f6e7334ae4fdcffd7e687a48d2f)

BUG=b:348506502
TEST=camera smoke test in brya

Change-Id: I6d95069f43af3cc9d7137235d7b730cfa1c6add3
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5793680
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-by: Hidenori Kobayashi <hidenorik@chromium.org>