CHROMIUM: sendmsg cros_net tp traces entry now
!!!DO NOT UPSTREAM!!!
!!!DO NOT PORT TO OTHER KERNELS!!!

Move the sendmsg tracepoint to capture the entry into inet_sendmsg.
This is needed so that eBPF programs attached to this tracepoint can
capture calling process context BEFORE packet routing occurs. Since eBPF
programs attached to kernel functions close to where packet routing
occurs cannot reliably retrieve process context; so the process
information on inet_sendmsg entry is captured, stored in a map and then
later retrieved by an eBPF program attached near where packet routing
occurs.

The reason why this patch shouldn't be upstreamed is that the only
reason these tracepoints are added is to provide an attach point for BPF
programs. Normally this would be done by hooking onto fentry/fexit.
Unfortunately fentry/fexit is not supported in the 5.10, 5.15 kernel for
ARM platforms. Porting over the upstream patches to add this
functionality to 5.10 and 5.15 also isn't feasible. A port attempt was
made and the number of patches was large, many conflicts needed to be
resolved and the changes touched a lot of the bpf subsystem and some
changes extended outside of the bpf subsystem.

So rather than making wide, sweeping and probably breaking changes to
the kernel; a relatively few trace points were added. Once all ARM
platforms move to a kernel that supports fentry/fexit tracepoints this
patchset should be dropped.

Ideally this patch would never make it over to the new kernel , but if
it does we can always manually drop it.

BUG=b:301276785
TEST=Ran tast.secagentd.networkevents.udp to make sure everything works
 on Atlas.
UPSTREAM-TASK=b:278303728

Signed-off-by: Jason Ling <jasonling@chromium.org>

Cq-Depend: chromium:4908436
Change-Id: Ifc1129079db39c0481f803a0db33eddac014bbf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4908623
Reviewed-by: Steven Rostedt <rostedt@google.com>
Commit-Queue: Jason Ling <jasonling@chromium.org>
Auto-Submit: Jason Ling <jasonling@chromium.org>
Tested-by: Jason Ling <jasonling@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>