UPSTREAM: media: ccs: Ensure control handlers have been set up after probe
If the sensor remains powered on after probe when it is needed again and
the runtime PM usage_count is incremented, the control handler setup for
neither control handler is run.

As this is, in most cases, a needless operation in probe, track the status
of sensor power handling after probe and set up control handlers even if
the device was already active right after probe.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
(cherry picked from commit a274f4d1e5af404fcc7b5554521913590c9d7b08)

BUG=b:348506502
TEST=camera smoke test in brya

Change-Id: Ie4c65e7db0cc2eca7462f707dcaac25c6f01f8c2
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5793683
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Hidenori Kobayashi <hidenorik@chromium.org>