CHROMIUM: Revert "Bluetooth: hci_core: Fix build warnings"
This reverts commit d9ce7d438366431e5688be98d8680336ce0a0f8d.

The reverted commit is known to cause crashes in chromeos-5.4 and
chromeos-4.19 and had to be reverted there during the most recent round
of stable release merges. Due to lack of CQ test coverage in chromeos-4.14,
it was not tested on hardware in this branch. Revert it out of caution.

The reverted commit was, from an operational perspective, completely
useless. The fixed build warning was not a real problem but imaginative,
assuming a BT interface index with more then four digits. Such interface
index values are not observed in the real world. Reverting this patch
has therefore no operational impact.

Note: This is a chromeos-4.14 only patch. There will be no upstreaming
task.

BUG=b:307318629
TEST=BT tests

Change-Id: Ic4f56e718a505026ca83bc50add8fd399a96e5c0
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4982162
Reviewed-by: Aashish Sharma <shraash@google.com>