FIXUP: CHROMIUM: hardlockup: detect hard lockups without NMIs using secondary cpus (match upstream format)
The format for the output for the buddy lockup detector changed a bit
(for the better) when it landed upstream. The format that landed
upstream also matches the format of the "perf" hardlockup
detector. Though there's not a huge advantage of backporting all of
the cleanups so we can get the exact upstream code, let's at least get
the format changes in our downstream version.

NOTE: this version uses trigger_cpumask_backtrace(), which is more
like how upstream worked prior to commit 1f38c86bb29f
("watchdog/hardlockup: avoid large stack frames in
watchdog_hardlockup_check()"). Doing it the upstream worked after that
commit also needs a bunch of extra backporting and the output here is
the same. In our case we use a static global to store the cpumask (to
avoid any large stackframe warnings) and that's not wasting too much
globals space because we never build kernels with a large
CONFIG_NR_CPUS.

DO NOT PICK THIS PATCH TO NEWER KERNELS DURING UPREVS. Newer kernels
will use the upstream code directly.

BUG=b:172213097
TEST=Format looks better

Change-Id: Ie6565d0b8f020e1b95bba4eb498ee5d45e533042
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5002610
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5003362