CHROMIUM: config: mediatek: Drop MDP3
The MDP3 device is no longer used. The MDP3 was previously used in
coordination with the decoder device to convert MM21 to NV12. Now that a
software detiler is being used in Chrome, there is no longer a user of
the MDP3.

CONFIG_VIDEO_MEDIATEK_ISP_DIP was removed automatically by running
./chromeos/scripts/kernelconfig olddefconfig

BUG=b:309657863
TEST=tast video tests on Cherry and Kukui
TEST=CQ

Signed-off-by: Nathan Hebert <nhebert@chromium.org>

Change-Id: I5e651dd9a2d83d1c99ed6d5a5afc5e5f653b75ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5008215
Tested-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>