CHROMIUM: config: mediatek: Drop MDP3
The MDP3 device is no longer used. The MDP3 was previously used in
coordination with the decoder device to convert MM21 to NV12. Now that a
software detiler is being used in Chrome, there is no longer a user of
the MDP3.

BUG=b:309657863
TEST=tast video tests on Asurada
TEST=Checked format with chromeos/scripts/kernelconfig olddefconfig
TEST=CQ

Signed-off-by: Nathan Hebert <nhebert@chromium.org>

Change-Id: I5e651dd9a2d83d1c99ed6d5a5afc5e5f653b75ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5014947
Tested-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Steve Cho <stevecho@chromium.org>