CHROMIUM: Block hibernate if Intel Keylocker is in use
Intel Keylocker stores the encryption key in SoC memory, this
memory is lost when hibernating to S5. The alternative would
be to hibernate to S4, which currently isn't well supported
(if at all) for Chrome OS devices.

The number of actual users of Intel Keylocker is supposed to
be very low, so let's block hibernate when Keylocker is in use.
With that the system can hibernate to S5 when Keylocker is not
used.

Check (again) if hibernation is available before creating the
hibernate snapshot, for if Keylocker use started after the
check in snapshot_open(). At that time userspace should be
frozen, so there shouldn't be any new Keylocker users after
that point (if hibernation is successful).

BUG=b:309149035
TEST=dd if=/dev/zero bs=100M count=1 of=/tmp/test.img
  cryptsetup luksFormat --cipher="capi:xts-aes-aeskl-plain" /tmp/test.img
  crosh> hibernate
    => the system eventually shuts down after multiple attempts to
       hibernate
UPSTREAM-TASK=b:309638461

Change-Id: I8424c4ed7f8d7e8fa0c745fdd873eb8c94d011bd
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5041747
Reviewed-by: Brian Geffon <bgeffon@chromium.org>