CHROMIUM: drm/evdi: Fix 2k monitor is not supported regression
This code introduces additional criteria when checking against an EDID
when the pixel clock exceeds capacity for supported resolution.

In the EDID, when there is no provided mode working under the supported
clock value the last criteria will test if there are other modes for
supported resolution and let it pass if it is the lowest available
frequency. It will be advertised as is, but trimmed to a lower frame
rate.

Example 1.: Supported modes <=4k@30 or <=2.5k@50
 __________________________
|  MODE   |SUPPORTED| PASS |
|_________|_________|______|
|4k@60    |    NO   |   NO |
|4k@30    |   YES   |  YES |
|2.5k@60  |    NO   |  YES |
|1080p@60 |   YES   |  YES |
|_________|_________|______|

System will be informed about the mode 2.5k@60 available, but this mode
will work under lower frequency, depending on the receiver capability

Example 2.: Supported modes <=2.5k@30
 __________________________
|  MODE   |SUPPORTED| PASS |
|_________|_________|______|
|4k@60    |    NO   |   NO |
|4k@30    |    NO   |   NO |
|2.5k@60  |    NO   |  YES |
|1080p@60 |   YES   |  YES |
|_________|_________|______|

BUG=b:296806688
TEST=tested on drallion kernel 4.19

Change-Id: I7a7815623b5e7ac1b84dbe2ddcc18278a0c2d933
Signed-off-by: Jakub Prussak <jakub.prussak@synaptics.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5036339
Reviewed-by: Gil Dekel <gildekel@chromium.org>