BACKPORT: arm64: dts: mediatek: mt8183: Change iospaces for thermal and svs
The SVS iospace starts at 0x1100bc00 and not at 0x1100b000 as the
latter is the thermal sensor iospace instead.

Change the iospaces for both as following:
 - Thermal: 0x1100b000, length 0xc00
 - SVS: 0x1100bc00, length 0x400

Please note that while this would be a breaking change for SVS (but
not for thermal sensors), it doesn't matter because the svs driver
never worked anyway because of the missing trips in tzts2, causing
that thermal zone to never actually register, hence the SVS driver
to fail probing anyway.

Link: https://lore.kernel.org/r/20231121125044.78642-2-angelogioacchino.delregno@collabora.com
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit e9ff6cdad87343447bf1a103068d91c0502e4629)

Conflicts:
   arch/arm64/boot/dts/mediatek/mt8183.dtsi
   The conflict is caused by https://lore.kernel.org/r/20240410083002.1357857-4-angelogioacchino.delregno@collabora.com

BUG=b:360282314
TEST=Check SVS is probed successfully

Change-Id: Id13add92620b72a3beaf4ba1677182dac810e637
Signed-off-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6055962
Reviewed-by: Pin-yen Lin <treapking@chromium.org>